+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Conditional rendering added.
 - Add a sup and sub notation.
+- Add the small notation.
 
 ## [4.1.0] - 2026-03-24